backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
Pillow==11.0.0
Werkzeug==3.1.3
PyPDF2==3.0.1
PyMuPDF==1.26.5
reportlab==4.0.7
pypdf==3.17.4
python-pptx==0.6.23
pytesseract==0.3.10
python-docx==1.1.0
